getForceDiff: Изменение типа short -> int8_t

+++ include/manipulator_control/ManipulatorControl.hpp
@@ -98,7 +98,7 @@ class ManipulatorControlHandler : public ManipulatorConfig<T> {
     ROS_DEBUG_STREAM(BLUE("needsForceControl_ = " << needsForceControl_));
     return needsForceControl_;
   };
-  short getForceDiff(const T force) const {
+  int8_t getForceDiff(const T force) const {
     const T forceDiff_{force - getTargetForce()};
     ROS_DEBUG_STREAM(BLUE("forceDiff_ = " << forceDiff_));
     if (needsForceControl(forceDiff_)) return -r2d2_math::sign(forceDiff_);
